Fix incorrect usage of update_or_create
Problem: fields in defaults like (firstName: None) were being used to update
models
Solution: don't pass all fields to defaults, prune the ones we want

The fields in the defaults keyword are used to update the existing model. So
we want to make sure that the defaults dict contains values that are valid.