Prevent expired tokens from granting access to api
In token authentication, we allow expired tokens to still make GET requests.
So we check if the request is a GET. If there is no request, we accept the
token. It turns out that the request is never passed as an argument to
`validate_token`, thus all expired tokens would grant access.

Rather than just pass the missing paramater, I changed the behavior to
disallow expired tokens from granting access.