Upgrade to caslib.py version (2.3.0) compatible with CAS 5
In that release, the profile returned from cas_oauth_client.get_profile
changed. Previously the profile was the raw response from the CAS api. In
order to be compatible with CAS 4 and CAS 5, caslib.py now hides the details
and just returns a predictable dict with profile attributes

Some duplicate error handling was removed here. In the
cas_oauth_client.get_profile call, any errors in the raw response raise
exceptions that we don't need to separately handle.